+++ punchbowl/level1/initial_uncertainty.py
@@ -93,6 +93,12 @@ def compute_uncertainty(data_array: np.ndarray,
     return noise
 
 
+def flag_saturated_pixels(data_object: NDCube) -> NDCube:
+    """Flag saturated pixels in the uncertainty layer."""
+    data_object.uncertainty.array[np.where(data_object.data >= data_object.meta["DSATVAL"].value)] = np.inf
+    return data_object
+
+
 @punch_task
 def update_initial_uncertainty_task(data_object: NDCube,
                                     bias_level: float = 100,
@@ -111,6 +117,8 @@ def update_initial_uncertainty_task(data_object: NDCube,
                                             bitrate_signal=bitrate_signal,
                                             )
     data_object.uncertainty = StdDevUncertainty(uncertainty_array)
+
+    data_object = flag_saturated_pixels(data_object)
 
     data_object.meta.history.add_now("LEVEL1-initial_uncertainty", "Initial uncertainty computed with:")
     data_object.meta.history.add_now("LEVEL1-initial_uncertainty", f"bias_level={bias_level}")